DATAMONGO-937 - Add support for creating text index.
We now support creating text index based on information gathered on domain types. 

Using @TextIndexed marks properties to be considered for the full text index. Use the weight attribute to influence document scoring during search operations. 

Please note that using @TextIndexed on entity properties forces all properties of any sub document to be considered as part of the text index. Any set weight will in that case be propagated to the siblings taking the most recent weight information into account, which means that a the weight attribute can be overridden for properties in sub documents.

The setting the index default language can be done via @Document(language) while @Language can be used to define the language_override field.

Still there has some cleanup to be done on the MongoPersistentEntityIndexResolver as there’s some duplicate code in there. Further on we’ve got to check this against mongodb 2.4 to verify test do not fail there - at least we’ve got to add a switch to disable tests in a 2.4 environment. 
The last thing still missing is proper reference documentation on text index.